same draft attempt 3
i think the issue was that when github checker or whatever opens this file, it opens it in Basic library. In EBeam library, the labels are in TEXT layer. But when you open it with Basic library, those labels end up in Layer 10/0? Because this gds file does pass the verification if EBeam library is chosen.